@@ -1,104 +1,64 @@
 {
   "name": "devorzo_backend",
-  "version": "0.1.3",
-  "description": "An open and awesome community for developers.",
-  "main": "dist/app.js",
+  "version": "0.2.0",
+  "description": "Backend devorzo",
+  "main": "app.js",
   "scripts": {
-    "start": "node dist/app.js --service=all --env=production",
+    "start": "node dist/app.js --env=production",
+    "dev": "ts-node-dev src/app.ts --env=development",
     "lint": "eslint src/**/*.{ts,js}",
     "lint:fix": "eslint src/**/*.{ts,js} --fix",
-    "prebuild": "echo prebuild",
     "build": "tsc -b",
-    "start:concur": "concurrently \"tsc -w\" \"nodemon dist\\app.js --service=all --env=development\"",
-    "start:dev": "nodemon dist/app.js --service=all --env=development",
     "watch": "tsc -w",
-    "pretest": "eslint --ignore-path .gitignore --ext ts,tsx,js,jsx src/*",
-    "test": "mocha -r ts-node/register tests/**/*.test.ts",
-    "coverage": "nyc -r lcov -e .ts -x \"*.test.ts\" npm run test"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/paritosh9199/major_project.git"
+    "url": "git+https://github.com/devorzo/devorzo_backend.git"
   },
-  "author": "Paritosh Srivastava",
-  "license": "ISC",
+  "author": "Paritosh",
+  "license": "AGPL-3.0",
   "bugs": {
-    "url": "https://github.com/paritosh9199/major_project/issues"
+    "url": "https://github.com/devorzo/devorzo_backend/issues"
   },
-  "homepage": "https://github.com/paritosh9199/major_project#readme",
+  "homepage": "https://github.com/devorzo/devorzo_backend#readme",
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "@types/chai": "^4.2.10",
-    "@types/connect-mongodb-session": "0.0.2",
-    "@types/cors": "^2.8.6",
-    "@types/ejs": "^3.0.1",
-    "@types/express": "^4.17.3",
-    "@types/express-session": "^1.17.0",
-    "@types/jsonwebtoken": "^8.3.8",
-    "@types/lodash": "^4.14.149",
+    "@commitlint/cli": "^9.1.2",
+    "@commitlint/config-conventional": "^9.1.2",
     "@types/minimist": "^1.2.0",
-    "@types/mocha": "^7.0.2",
-    "@types/mongoose": "^5.7.6",
-    "@types/multer": "^1.4.2",
-    "@types/multer-gridfs-storage": "^4.0.0",
-    "@types/node": "^13.9.0",
-    "@types/node-fetch": "^2.5.6",
-    "@types/uuid": "^7.0.0",
-    "@types/validator": "^12.0.1",
-    "@typescript-eslint/eslint-plugin": "^2.23.0",
-    "@typescript-eslint/parser": "^2.23.0",
-    "assert": "^2.0.0",
-    "chai": "^4.2.0",
-    "concurrently": "^5.1.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.1",
+    "@types/mongoose": "^5.7.36",
+    "@types/uuid": "^8.3.0",
+    "@typescript-eslint/eslint-plugin": "^3.9.1",
+    "@typescript-eslint/parser": "^3.9.1",
+    "eslint": "^7.7.0",
+    "eslint-config-airbnb-base": "^14.2.0",
+    "eslint-plugin-import": "^2.22.0",
     "husky": "^4.2.5",
-    "lint-staged": "^10.2.11",
-    "mocha": "^7.1.0",
-    "nodemon": "^2.0.2",
-    "nyc": "^15.0.0",
-    "ts-node": "^8.6.2"
+    "nodemon": "^2.0.4",
+    "ts-node-dev": "^1.0.0-pre.58"
   },
   "dependencies": {
-    "@sendgrid/mail": "^7.1.0",
-    "@types/helmet": "0.0.47",
-    "axios": "^0.19.2",
-    "bcryptjs": "^2.4.3",
+    "@typegoose/typegoose": "^7.3.2",
+    "apollo-server-express": "^2.16.1",
     "body-parser": "^1.19.0",
-    "connect-mongo": "^3.2.0",
-    "connect-mongodb-session": "^2.3.1",
+    "class-validator": "^0.12.2",
     "cors": "^2.8.5",
-    "domino": "^2.1.4",
-    "ejs": "^3.0.1",
     "express": "^4.17.1",
-    "express-session": "^1.17.0",
-    "helmet": "^3.22.0",
-    "jsonwebtoken": "^8.5.1",
-    "lodash": "^4.17.15",
-    "minimist": "^1.2.4",
-    "mongoose": "^5.9.13",
-    "multer": "^1.4.2",
-    "multer-gridfs-storage": "^4.0.2",
-    "node-fetch": "^2.6.0",
-    "page-metadata-parser": "^1.1.4",
-    "path": "^0.12.7",
-    "typescript": "^3.8.3",
-    "uuid": "^7.0.2",
-    "validator": "^12.2.0"
-  },
-  "lint-staged": {
-    "src/**/*.{ts,js}": [
-      "npm run lint:fix",
-      "git add"
-    ]
+    "graphql": "^15.3.0",
+    "helmet": "^4.1.0",
+    "minimist": "^1.2.5",
+    "mongodb": "^3.6.0",
+    "mongoose": "^5.10.0",
+    "reflect-metadata": "^0.1.13",
+    "type-graphql": "^1.0.0-rc.3",
+    "typescript": "^3.9.7",
+    "uuid": "^8.3.0",
+    "validator": "^13.1.1"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged"
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "yarn lint:fix"
     }
   }
 }
