@@ -1,9 +1,10 @@
-export const  generateHexString = function (len = 8) {
+export const generateHexString = (len = 8): string => {
+  const hex = '0123456789ABCDEF';
+  let output = '';
+  for (let i = 0; i < len; ++i) {
+    output += hex.charAt(Math.floor(Math.random() * hex.length));
+  }
+  return output;
+};
 
-    const hex = "0123456789ABCDEF"
-    let output = ""
-    for (let i = 0; i < len; ++i) {
-        output += hex.charAt(Math.floor(Math.random() * hex.length))
-    }
-    return output
-}
+export default { generateHexString };
